Add focused test runner that builds and cleans up static assets
The test suite reads the bundled static client assets (src/ui/static/*.js)
at import time (src/features/assets.ts), and the Stripe tests need stripe-mock
running with STRIPE_MOCK_HOST/PORT set. Previously only the full runner
(`deno task test`) performed this setup, so running a subset of files directly
with `deno test` failed on a missing admin.js, or called the real Stripe API
and returned an error page instead of a checkout redirect.

Extract the shared setup/teardown into scripts/test-harness.ts (stripe-mock
lifecycle, static-asset build, and the `deno test` invocation) and add a
`deno task test:files` entry point that reuses it to run only the requested
files. Both runners now build the static assets up front and remove any they
generated afterwards, restoring a fresh checkout to its pristine state while
preserving assets a developer built themselves.

- build-static-assets.ts: export STATIC_ASSET_OUTFILES as the single source
  of truth for generated outputs, used by the harness for cleanup
- run-tests.ts: slim down to a thin entry point over the shared harness
- CLAUDE.md: document test:files as the way to run individual files
